Déplacer les fonctions concernant les activités méthodes transports de manière générale dans un fichier à part de dossier
